+++ src/State/PlaylistScreenRegionProvider.php
@@ -72,7 +72,7 @@ final class PlaylistScreenRegionProvider extends AbstractProvider
         $output->id = $object->getId();
         $output->playlist = $this->playlistProvider->toOutput($object->getPlaylist());
         $output->weight = $object->getWeight();
-        $output->setRelationsModified($object->getRelationsModified());
+        $output->setRelationsChecksum($object->getRelationsChecksum());
 
         return $output;
     }