+++ CHANGELOG.md
@@ -15,6 +15,10 @@ All notable changes to this project will be documented in this file.
   - Removed PHP Upgrade coding standards github actions check.
   - Changed user identifier from email to providerId. Made email nullable. Copied value from email to providerId in
 migration.
+- [#184](https://github.com/os2display/display-api-service/pull/184)
+  - Added RelationsModifiedTrait to serialization groups.
+- [#186](https://github.com/os2display/display-api-service/pull/186)
+  - Fix for "relations modified" not set correctly on OneToMany relations
 - [#185](https://github.com/os2display/display-api-service/pull/185)
   - Disable RelationsModified listener when loading fixtures to optimize performance
 - [#182](https://github.com/os2display/display-api-service/pull/182)